+++ CHANGELOG.md
@@ -14,6 +14,7 @@ OpenNMT-tf follows [semantic versioning 2.0.0](https://semver.org/). The API cov
 
 ### New features
 
+* Average attention network ([Zhang et al. 2018](https://arxiv.org/abs/1805.00631)) in the Transformer decoder
 * `--log_prediction_time` flag to summarize inference execution time:
   * total prediction time
   * average prediction time